06-api-integration/fastapi-app/requirements.txt
# ─── Core Framework ──────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.0
pydantic-settings==2.5.2

# ─── Authentication & JWT ────────────────────────────────────
python-jose[cryptography]==3.3.0
# python-jose with cryptography backend for RS256 support

# ─── HTTP Client (for JWKS fetching) ────────────────────────
httpx==0.27.2

# ─── Environment Variables ───────────────────────────────────
python-dotenv==1.0.1

# ─── CORS & Security ────────────────────────────────────────
# CORS is built into FastAPI middleware

# ─── Logging ─────────────────────────────────────────────────
python-json-logger==2.0.7

# ─── Rate Limiting ───────────────────────────────────────────
slowapi==0.1.9

# ─── Development Dependencies ───────────────────────────────
# Install these with: pip install -r requirements-dev.txt
# pytest==8.3.0
# pytest-asyncio==0.24.0
# httpx==0.27.2  (also used for test client)
# coverage==7.6.0
